port:parallel: Replace int with unsigned long to fix build error
Reference
http://photonics.engr.uga.edu/piklab/index.html
I do not know if this has any side effects when parallel port programmers
are used. Perhaps the programmers/maintainers take notice.
They never responded to my messages, unfortunately.

+++ src/common/port/parallel.h
@@ -52,7 +52,7 @@ private:
   int     _fd;
   QString _device;
   struct RequestData {
-    int read, write;
+    unsigned long read, write;
   };
   static const RequestData REQUEST_DATA[Nb_RequestTypes];
   uchar _images[Nb_RequestTypes];
